Move query-related tests to `query_test.go`.
